+++ src/console.rs
@@ -6,6 +6,10 @@ pub fn info(message: &str) {
     println!("{}", NotSet.bold().paint(message));
 }
 
+pub fn warn(message: &str) {
+    println!("{}", Yellow.bold().paint(message));
+}
+
 pub fn success(message: &str) {
     println!("{}", Green.bold().paint(message));
 }